@@ -224,7 +224,7 @@ setup(
     },
     long_description=open('README.md').read(),
     long_description_content_type="text/markdown",
-    python_requires='>=3.6',
+    python_requires='>=3.7',
     install_requires=[
         "osqp >= 0.4.1",
         "ecos >= 2",
